Seperated RecValEnumerator from Enumerable class
